requirements.txt
streamlit
tensorflow
opencv-python
pillow
numpy
huggingface_hub
uvicorn
fastapi
streamlit_lottie